+++ ScoopItFramework/Classes/Model/SISource.h
@@ -11,7 +11,7 @@
 
 
 @interface SISource : SIModel {
-	int lid;
+	long long lid;
 	NSString* name;
 	NSString* description;
 	NSString* type;
@@ -19,7 +19,7 @@
 	NSString* url;
 }
 
-@property (nonatomic) int lid;
+@property (nonatomic) long long lid;
 @property (nonatomic, retain) NSString* name;
 @property (nonatomic, retain) NSString* description;
 @property (nonatomic, retain) NSString* type;
